Extract retrievers from the provider
