+++ aztokenprovider/retriever_msi.go
@@ -1,0 +1,60 @@
+package aztokenprovider
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/grafana/grafana-azure-sdk-go/azcredentials"
+	"github.com/grafana/grafana-azure-sdk-go/azsettings"
+)
+
+type managedIdentityTokenRetriever struct {
+	clientId   string
+	credential azcore.TokenCredential
+}
+
+func getManagedIdentityTokenRetriever(settings *azsettings.AzureSettings, credentials *azcredentials.AzureManagedIdentityCredentials) TokenRetriever {
+	var clientId string
+	if credentials.ClientId != "" {
+		clientId = credentials.ClientId
+	} else {
+		clientId = settings.ManagedIdentityClientId
+	}
+	return &managedIdentityTokenRetriever{
+		clientId: clientId,
+	}
+}
+
+func (c *managedIdentityTokenRetriever) GetCacheKey() string {
+	clientId := c.clientId
+	if clientId == "" {
+		clientId = "system"
+	}
+	return fmt.Sprintf("azure|msi|%s", clientId)
+}
+
+func (c *managedIdentityTokenRetriever) Init() error {
+	options := &azidentity.ManagedIdentityCredentialOptions{}
+	if c.clientId != "" {
+		options.ID = azidentity.ClientID(c.clientId)
+	}
+	credential, err := azidentity.NewManagedIdentityCredential(options)
+	if err != nil {
+		return err
+	} else {
+		c.credential = credential
+		return nil
+	}
+}
+
+func (c *managedIdentityTokenRetriever) GetAccessToken(ctx context.Context, scopes []string) (*AccessToken, error) {
+	accessToken, err := c.credential.GetToken(ctx, policy.TokenRequestOptions{Scopes: scopes})
+	if err != nil {
+		return nil, err
+	}
+
+	return &AccessToken{Token: accessToken.Token, ExpiresOn: accessToken.ExpiresOn}, nil
+}
